requirements.txt
atram==2.1
atomicwrites==1.2.1
attrs==18.2.0
biopython==1.72
coverage==4.5.1
hypothesis==3.76.0
more-itertools==4.3.0
numpy==1.15.2
pluggy==0.7.1
psutil==5.4.7
py==1.7.0
pytest==3.8.2
six==1.11.0
